app/customers/admin.py
from django.contrib import admin
from . import models


class CustomerAdmin(admin.ModelAdmin):
    list_display = ('name', 'email', 'phone', 'created_at')
    search_fields = ('name', 'email', 'phone')
    
    
    
admin.site.register(models.Customer, CustomerAdmin)